+++ src/OpenSSL/crypto.py
@@ -24,7 +24,7 @@ elif sys.version_info < (3, 8):
 else:
     from typing_extensions import deprecated
 
-from cryptography import x509
+from cryptography import utils, x509
 from cryptography.hazmat.primitives.asymmetric import (
     dsa,
     ec,
@@ -272,7 +272,7 @@ class PKey:
             der = dump_publickey(FILETYPE_ASN1, self)
             return typing.cast(_Key, load_der_public_key(der))
         else:
-            der = dump_privatekey(FILETYPE_ASN1, self)
+            der = _dump_privatekey_internal(FILETYPE_ASN1, self)
             return typing.cast(_Key, load_der_private_key(der, password=None))
 
     @classmethod
@@ -333,6 +333,10 @@ class PKey:
             )
             return load_privatekey(FILETYPE_ASN1, der)
 
+    @deprecated(
+        "PKey.generate_key is deprecated. You should use the key "
+        "generation APIs in cryptography instead."
+    )
     def generate_key(self, type: int, bits: int) -> None:
         """
         Generate a key pair of the given type, with the given number of bits.
@@ -389,6 +393,10 @@ class PKey:
 
         self._initialized = True
 
+    @deprecated(
+        "PKey.check is deprecated. You should use the APIs in "
+        "cryptography instead."
+    )
     def check(self) -> bool:
         """
         Check the consistency of an RSA private key.
@@ -825,6 +833,10 @@ class X509:
         der = crypto_cert.public_bytes(Encoding.DER)
         return load_certificate(FILETYPE_ASN1, der)
 
+    @deprecated(
+        "X509.set_version is deprecated. You should use "
+        "cryptography's CertificateBuilder instead."
+    )
     def set_version(self, version: int) -> None:
         """
         Set the version number of the certificate. Note that the
@@ -864,6 +876,10 @@ class X509:
         pkey._only_public = True
         return pkey
 
+    @deprecated(
+        "X509.set_pubkey is deprecated. You should use "
+        "cryptography's CertificateBuilder instead."
+    )
     def set_pubkey(self, pkey: PKey) -> None:
         """
         Set the public key of the certificate.
@@ -879,6 +895,10 @@ class X509:
         set_result = _lib.X509_set_pubkey(self._x509, pkey._pkey)
         _openssl_assert(set_result == 1)
 
+    @deprecated(
+        "X509.sign is deprecated. You should use "
+        "cryptography's CertificateBuilder instead."
+    )
     def sign(self, pkey: PKey, digest: str) -> None:
         """
         Sign the certificate with this key and digest type.
@@ -966,6 +986,10 @@ class X509:
         """
         return _lib.X509_subject_name_hash(self._x509)
 
+    @deprecated(
+        "X509.set_serial_number is deprecated. You should use "
+        "cryptography's CertificateBuilder instead."
+    )
     def set_serial_number(self, serial: int) -> None:
         """
         Set the serial number of the certificate.
@@ -1014,6 +1038,10 @@ class X509:
         finally:
             _lib.BN_free(bignum_serial)
 
+    @deprecated(
+        "X509.gmtime_adj_notAfter is deprecated. You should use "
+        "cryptography's CertificateBuilder instead."
+    )
     def gmtime_adj_notAfter(self, amount: int) -> None:
         """
         Adjust the time stamp on which the certificate stops being valid.
@@ -1028,6 +1056,10 @@ class X509:
         notAfter = _lib.X509_getm_notAfter(self._x509)
         _lib.X509_gmtime_adj(notAfter, amount)
 
+    @deprecated(
+        "X509.gmtime_adj_notBefore is deprecated. You should use "
+        "cryptography's CertificateBuilder instead."
+    )
     def gmtime_adj_notBefore(self, amount: int) -> None:
         """
         Adjust the timestamp on which the certificate starts being valid.
@@ -1079,6 +1111,10 @@ class X509:
     ) -> None:
         return _set_asn1_time(which(self._x509), when)
 
+    @deprecated(
+        "X509.set_notBefore is deprecated. You should use "
+        "cryptography's CertificateBuilder instead."
+    )
     def set_notBefore(self, when: bytes) -> None:
         """
         Set the timestamp at which the certificate starts being valid.
@@ -1105,6 +1141,10 @@ class X509:
         """
         return self._get_boundary_time(_lib.X509_getm_notAfter)
 
+    @deprecated(
+        "X509.set_notAfter is deprecated. You should use "
+        "cryptography's CertificateBuilder instead."
+    )
     def set_notAfter(self, when: bytes) -> None:
         """
         Set the timestamp at which the certificate stops being valid.
@@ -1151,6 +1191,10 @@ class X509:
         self._issuer_invalidator.add(name)
         return name
 
+    @deprecated(
+        "X509.set_issuer is deprecated. You should use "
+        "cryptography's CertificateBuilder instead."
+    )
     def set_issuer(self, issuer: X509Name) -> None:
         """
         Set the issuer of this certificate.
@@ -1179,6 +1223,10 @@ class X509:
         self._subject_invalidator.add(name)
         return name
 
+    @deprecated(
+        "X509.set_subject is deprecated. You should use "
+        "cryptography's CertificateBuilder instead."
+    )
     def set_subject(self, subject: X509Name) -> None:
         """
         Set the subject of this certificate.
@@ -1673,6 +1721,10 @@ def dump_privatekey(
 
     :return: The buffer with the dumped key in
     :rtype: bytes
+
+    .. deprecated:: 26.3.0
+       Use the serialization APIs on ``cryptography`` private key types
+       instead.
     """
     bio = _new_mem_buf()
 
@@ -1720,6 +1772,20 @@ def dump_privatekey(
     _openssl_assert(result_code != 0)
 
     return _bio_to_string(bio)
+
+
+_dump_privatekey_internal = dump_privatekey
+
+utils.deprecated(
+    dump_privatekey,
+    __name__,
+    (
+        "dump_privatekey is deprecated. You should use the APIs in "
+        "cryptography."
+    ),
+    DeprecationWarning,
+    name="dump_privatekey",
+)
 
 
 class _PassphraseHelper: